Dummy program copies a CSV file and adds a sum as the final column
This is just the starting point for a program that can read in a CSV file
and build a NetCDF file that can be read by the hierarchical clustering
program. I am checking it in because it's in a working state I may want
to revert to if things go wrong as I proceed.