Successfully read and cluster aircraft data.
Reads in the file created by csv_to_netcdf.F90 and clusters the data and
saves it out. There are number of shortcomings right now. First, the file
contains NaN signifiers (-99999) and those are not handled yet. They get
assigned a R of 0 because they have 0 variance, so they are last the cluster
anyway. Second, there's no date associated with the files so the output
files get saved with a bunch of **** in them. Third, this code is in
desperate need of refactoring. There are now nested logic branches for
whether we're using aircraft data, whether we're reading/saving the
dissimilarity matrix, and whether we've got a good MPI setup for tiling
the dissimilarity matrix (for more efficient MPI distribution).