Fixed an MPI_ALLGATHER bug
MPI_ALLGATHER was not behaving well when combining the MPI_REAL and the
MPI_TYPE_CONTIGUOUS nodeType, even though they should have been the same
and it worked on U1 and for some situations on U2. The call is now
easier to read and seems to work for all cases on U2.

Also got rid of a bunch of debug printing.