Basic checkpointing works
Code is real ugly and need cleaned up and for checkpoint
timestep to be paramatrized isntead of hardcoded. But I
did a test with 384x384 grid points and the restarted
version reproduces the original clustering.

printCheckpointFile doesn't work on big files but I'm no
sure what the fail point is or why.