Add exit-after-checkpoint option, fix bug
When the program times out on the cluster, it leaves a mess of files
everywhere, so I added an option to exit after a checkpoint has finished
that way it cleans up its files and you can read the output in the
listings. The default is for this behaviour to not happen if it's not
specified in the arglist.lst file.

I also fixed a small bug with non-square input. GRID_NI and GRID_NJ
were swapped in the indexing into the input buffer from the netcdf
files.