Move checkpoint load before dissmatrix
The way I initially implemented the checkpoint loading, it wasn't
happening until after the dissimilarity matrix checks, which
was super inefficient since it was loading up the dissimilarity
matrix, filling the priority queues and then blowing all that
away to load the checkpoint file. Now the checkpoint file is
loaded up before that and it skips the dissimilarity matrix
if there was a checkpoint. I had to move some initializations
to up before the checkpoint code to make this work.